fix ctrl click thread
