+++ libs/components/src/lib/components/ThreadListChannel/index.tsx
@@ -1,26 +1,15 @@
-import { useMenu } from '@mezon/core';
 import {
-	appActions,
-	referencesActions,
 	selectAllThreadUnreadBehind,
 	selectCategoryExpandStateByCategoryId,
 	selectChannelMetaById,
 	selectChannelMetaEntities,
-	selectCloseMenu,
 	selectCurrentChannelId,
-	threadsActions,
-	useAppDispatch,
 	useAppSelector
 } from '@mezon/store';
 import type { IChannel } from '@mezon/utils';
 import React, { useMemo } from 'react';
 import { useSelector } from 'react-redux';
 import ThreadLink from './ThreadLink';
-
-type ThreadListChannelProps = {
-	threads: IChannel[];
-	isCollapsed: boolean;
-};
 
 export type ListThreadChannelRef = {
 	scrollIntoThread: (threadId: string, options?: ScrollIntoViewOptions) => void;
@@ -38,23 +27,6 @@ export const ThreadLinkWrapper: React.FC<ThreadLinkWrapperProps> = ({ thread, no
 	const isCategoryExpanded = useAppSelector((state) => selectCategoryExpandStateByCategoryId(state, thread.category_id as string));
 	const allThreadBehind = useAppSelector((state) => selectAllThreadUnreadBehind(state, thread?.clan_id, thread?.parent_id, thread?.id));
 	const channelMetadata = useSelector(selectChannelMetaEntities);
-
-	const closeMenu = useAppSelector(selectCloseMenu);
-	const dispatch = useAppDispatch();
-	const { setStatusMenu } = useMenu();
-
-	const handleClickLink = (thread: IChannel) => {
-		dispatch(referencesActions.setOpenEditMessageState(false));
-		if (currentChannelId === thread.parent_id) {
-			dispatch(threadsActions.setIsShowCreateThread({ channelId: thread.parent_id as string, isShowCreateThread: false }));
-		}
-		if (closeMenu) {
-			setStatusMenu(false);
-		}
-		dispatch(threadsActions.setOpenThreadMessageState(false));
-		dispatch(threadsActions.setValueThread(null));
-		dispatch(appActions.setIsShowCanvas(false));
-	};
 
 	const isShowThread = (thread: IChannel) => {
 		return (
@@ -84,5 +56,5 @@ export const ThreadLinkWrapper: React.FC<ThreadLinkWrapperProps> = ({ thread, no
 		return null;
 	}
 
-	return <ThreadLink isActive={isActive} thread={thread} hasLine={hasUnreadThreadBehind} handleClick={handleClickLink} />;
+	return <ThreadLink isActive={isActive} thread={thread} hasLine={hasUnreadThreadBehind} currentChannelId={currentChannelId as string} />;
 };
